Created Interface ItemClickListener
Created Interface ItemClickListener, add method onCLick(View v, Int position).

Then in CustomAdpater,  firstly, we created the variable of type Interface and named as clickListener, then we created the method setClickListener(ItemClickListener mylistener) intialized the clickListener variable, then in Myviewholder implements the View.OnClickListener, override the onClick(), then in Onclick() we are checking if clickListener is null if it is then we are calling the method of or interface, setting the v, getadapterPosition(). Then in mainactivity, we are implementing the ItemClickListenerInterface, overriding our onCLIck() method, we are setting our adapter by calling setClickListener(this).